Revert "Move all abstract tests into a single file."
This reverts commit f668782eb9f727f6b58fdba4f0fd9dd9f58eb382.
